backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.0
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
ccxt==4.1.22
websockets==12.0
python-multipart==0.0.6